ENH: Add itk::DefaultParameterValueType
ITK was hard coded to double precision for transforms
and optimizers, and registration could only be completed
in double precision.  This provides a single
place where the default type can be changed to single
precision for testing purposes so that it is easier
to determine where there are potiential type mis-matches
and where templates need to be introduced.

Change-Id: Iade06bae738e512f3b7fd2e0d0a1aed085be7366